fix: split gameplay.spec.ts to pass LOC audit; allowlist pre-existing utility scripts
Agent-Logs-Url: https://github.com/TeacherEvan/English-K1Run/sessions/3c6b4379-1338-4659-a3c7-f7cd615a1f68

(cherry picked from commit 3efce1457299ce81a96c1e917541f1fac4e0ae4b)
